fix: 修复类型转换错误 Co-authored-by: monkeycode-ai <monkeycode-ai@chaitin.com>

+++ Assets/Scripts/Manager/ArchiveManager.cs
@@ -1400,7 +1400,7 @@ public class ArchiveManager : CommonInstance<ArchiveManager>
         
         // 添加基础技能（根据元素类型）
         SingleSkillData singleSkill = new SingleSkillData();
-        singleSkill.skillId = BattleManager.Instance.PuGongIdByYuanSu((YuanSuType)p.yuanSu);
+        singleSkill.skillId = (int)BattleManager.Instance.PuGongIdByYuanSu((YuanSuType)p.yuanSu);
         singleSkill.skillLevel = 1;
         p.allSkillData.skillList.Add(singleSkill);
         p.allSkillData.equippedSkillIdList.Add(singleSkill.skillId);
